requirements-dev.txt
pytest==8.3.4
pytest-asyncio==0.25.0
pydantic==2.10.4
PyYAML==6.0.2
httpx==0.28.1
fastapi==0.115.6
Jinja2==3.1.5
ruff==0.8.4

